合并数据

See merge request !12

+++ src/main/java/com/shsnc/base/authorization/bean/AuthorizationRole.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
  */
 public class AuthorizationRole implements Serializable {
 
-    private Integer roleId;
+    private Long roleId;
 
     private String roleName;
 
@@ -19,11 +19,11 @@ public class AuthorizationRole implements Serializable {
 
     private Integer orders;
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
