+++ src/main/java/com/shsnc/base/authorization/model/AuthorizationInfoModel.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
  */
 public class AuthorizationInfoModel implements Serializable{
 
-    private Integer authorizationId;
+    private Long authorizationId;
 
     private String authorizationName;
 
@@ -39,11 +39,11 @@ public class AuthorizationInfoModel implements Serializable{
         }
     }
 
-    public Integer getAuthorizationId() {
+    public Long getAuthorizationId() {
         return authorizationId;
     }
 
-    public void setAuthorizationId(Integer authorizationId) {
+    public void setAuthorizationId(Long authorizationId) {
         this.authorizationId = authorizationId;
     }
